rename manifest-migrate -> manifest-storage-layout
