+++ packages/core/src/manifest-storage-layout.ts
